+++ l4/l4_e1/src/main.c
@@ -39,7 +39,7 @@ static K_SEM_DEFINE(run_app, 0, 1);
 /* STEP 10.1 - Declare the variable to store the client ID */
 
 
-static void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 				   struct net_if *iface)
 {
 	if ((mgmt_event & EVENT_MASK) != mgmt_event) {
